move yaml endpoints
